requirements.txt
beautifulsoup4==4.10.0
codecarbon==2.6.0
huggingface_hub==0.23.4
laserembeddings==1.1.2
matplotlib==3.9.2
mysql_connector_repackaged==0.3.1
numpy==1.26.4
pandas==2.2.2
plotly==5.20.0
scikit_learn==1.5.1
seaborn==0.13.2
sentence_transformers==3.0.1
SQLAlchemy==2.0.32
streamlit==1.38.0
torch==1.13.1
tqdm==4.66.5
transformers==4.41.2
fasttext
sentencepiece
pipedream